refactor(core): move `MapPhraseLinter` to its own module
