+++ harper-core/src/linting/mod.rs
@@ -26,6 +26,7 @@ mod lint;
 mod lint_group;
 mod lint_kind;
 mod long_sentences;
+mod map_phrase_linter;
 mod matcher;
 mod merge_linters;
 mod merge_words;
@@ -92,6 +93,7 @@ pub use lint::Lint;
 pub use lint_group::{LintGroup, LintGroupConfig};
 pub use lint_kind::LintKind;
 pub use long_sentences::LongSentences;
+pub use map_phrase_linter::MapPhraseLinter;
 pub use matcher::Matcher;
 pub use merge_words::MergeWords;
 pub use multiple_sequential_pronouns::MultipleSequentialPronouns;
